requirements_old.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anaconda-anon-usage @ file:///home/task_175762568762291/conda-bld/anaconda-anon-usage_1757625705334/work
anaconda-auth @ file:///home/task_175994313587130/conda-bld/anaconda-cloud-auth-split_1759943195502/work
anaconda-cli-base @ file:///home/task_176069795959605/conda-bld/anaconda-cli-base_1760698232645/work
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
anthropic==0.72.0
anyio==4.11.0
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
attrs==25.4.0
backtesting==0.6.5
bitarray==3.7.2
bokeh==3.8.0
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==6.2.1
ccxt==4.5.14
certifi @ file:///home/task_175978569032317/conda-bld/certifi_1759786061720/work/certifi
cffi @ file:///home/task_176009805548162/conda-bld/cffi_1760098067589/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
ckzg==2.1.5
click @ file:///croot/click_1744271578095/work
conda @ file:///home/task_176045791080537/conda-bld/conda_1760457922380/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1755123313635/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
contourpy==1.3.3
cryptography @ file:///home/task_176010605113983/conda-bld/cryptography-split_1760106309120/work
cytoolz==1.1.0
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
docstring_parser==0.17.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
groq==0.33.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperliquid-python-sdk==0.20.0
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
jaraco.classes @ file:///croot/jaraco.classes_1755516327832/work
jaraco.context @ file:///io/perseverance-python-buildout/croot/jaraco.context_1731706750665/work
jaraco.functools @ file:///croot/jaraco.functools_1740408111753/work
jeepney @ file:///tmp/build/80754af9/jeepney_1627537048313/work
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer @ file:///croot/jsonpointer_1753788433671/work
keyring @ file:///croot/keyring_1745484691218/work
libmambapy @ file:///home/task_175762302885340/conda-bld/mamba-split_1757623046707/work/libmambapy
llvmlite==0.44.0
markdown-it-py @ file:///croot/markdown-it-py_1756299186942/work
MarkupSafe==3.0.3
mdurl @ file:///home/task_175855216352428/conda-bld/mdurl_1758552176751/work
menuinst @ file:///croot/menuinst_1753464582829/work
more-itertools @ file:///home/task_176112148319406/conda-bld/more-itertools_1761121494880/work
msgpack==1.1.2
multidict==6.7.0
narwhals==2.10.1
numba==0.61.2
numpy==2.2.6
openai==2.6.1
packaging @ file:///home/task_176104878999627/conda-bld/packaging_1761049078606/work
pandas==2.3.3
pandas-ta==0.4.71b0
parsimonious==0.10.0
pillow==12.0.0
pkce @ file:///work/perseverance-python-buildout/croot/pkce_1728388093137/work
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///home/task_175749564936714/conda-bld/pycparser_1757496040682/work
pycryptodome==3.23.0
pydantic @ file:///home/task_176069749099755/conda-bld/pydantic_1760697502973/work
pydantic-settings @ file:///home/task_175878416925449/conda-bld/pydantic-settings_1758784193926/work
pydantic_core @ file:///home/task_176059738185733/conda-bld/pydantic-core_1760597433577/work
Pygments @ file:///croot/pygments_1744664109463/work
PyJWT @ file:///io/buildout/croot/pyjwt_1739558969999/work
pyparsing==3.2.5
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
pytz==2025.2
PyYAML==6.0.3
readchar @ file:///home/task_176061336965958/conda-bld/readchar_1760613437376/work
regex==2025.10.23
requests @ file:///croot/requests_1756709366904/work
rich @ file:///home/task_176037557612320/conda-bld/rich_1760375632000/work
rlp==4.1.0
rsa==4.9.1
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
scikit-learn==1.7.2
scipy==1.16.3
SecretStorage @ file:///home/task_175793143758247/conda-bld/secretstorage_1757931467244/work
semver @ file:///croot/semver_1736540782504/work
setuptools==80.9.0
shellingham @ file:///work/perseverance-python-buildout/croot/shellingham_1728395303315/work
six==1.17.0
sniffio==1.3.1
ta==0.11.0
termcolor==3.2.0
threadpoolctl==3.6.0
tomli @ file:///croot/tomli_1753774587605/work
toolz==1.1.0
tornado==6.5.2
tqdm @ file:///croot/tqdm_1738943501192/work
truststore @ file:///croot/truststore_1753907171582/work
typer @ file:///home/task_175758067631213/conda-bld/typer_1757580709537/work
typing-inspection @ file:///home/task_176061348700415/conda-bld/typing-inspection_1760614166875/work
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.2
uritemplate==4.2.0
urllib3 @ file:///croot/urllib3_1750775463400/work
websocket-client==1.9.0
wheel==0.45.1
xyzservices==2025.10.0
yarl==1.22.0
zstandard @ file:///home/task_175818846949212/conda-bld/zstandard_1758189073015/work
solders==0.27.0
jsonalias==0.1.1
